Stricter pinning in Cargo.toml
Because we rely on a modified `pipeline-cache.rs` copied directly from
the Bevy codebase, we need to pin as closely as possible to the `wgpu`
and `naga` versions that were used at the time Bevy wrote the
`pipeline_cache.rs` file.

It might also be worth using `Cargo.lock`, as that is now recommended by
the Cargo Team themselves: https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html

Longer-term it may even be worth re-architecting the codebase to follow
Bevy's `gpu_readback.rs` example that copies Render World data to the
App World using a channel. This would we wouldn't even need the
`pipeline_cache.rs` file.

Address #8